Initial version of gitlab mr validator
Building a trivial version of a gitlab mr validator, which is checking
that commits associated to a gitlab merge request are matching the
current sevdesk commit conventions.

The convention states, that each commit needs a conventional commit
prefix followed by the ticket number and a short title. The body
of the commit message must state the why: Describing why the changes
where made, the implications and possible issues.

The purpose of the tool, is to automate some bits and pieces regarding
code reviews. It's possible to extend the validator to auto-close
merge requests that are not matching the target format.