Implemented deny of PR's
Now the bot will also close a merge request if it's invalid. But to ensure that the author of the mr knows what the reason was, the bot will post a message indicating that the git commits where not matching the coding conventions.

This applies only to mrs which are not in draft state. For draft MR's only a warning will be logged containing the mr id.